farmpower_backend_v2/requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.10
python-socketio==5.11.2
python-engineio==4.9.1
pydantic==2.8.2
python-dotenv==1.0.1
